split factories to reduce obj file sections
