+++ source/engine/engine/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
         ComponentFactories.cpp
+        ModuleFactory.cpp
         NcEngineImpl.cpp
         RegistryFactories.cpp
 )
